requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.25
python-dotenv>=1.0
pydantic>=2.6
stripe>=8.0
anthropic>=0.40
